refactor(mcp): isolate result handle shaping
