+++ crates/coral-mcp/src/server.rs
@@ -7,13 +7,8 @@ use coral_api::v1::{
     TableSummary as ProtoTableSummary, catalog_item,
 };
 use coral_client::{
-    AppClient, CatalogClient, CollectedQueryResult, FeedbackClient, QueryClient, QueryResultError,
-    SourceClient, batches_to_json_rows_json_safe_numbers, decode_execute_sql_response,
-    default_workspace,
-    result_slice::{
-        ColumnSummary, ResultPage, ResultSliceRequest, has_duplicate_column_names,
-        result_estimated_bytes, schema_summary, slice_result,
-    },
+    AppClient, CatalogClient, CollectedQueryResult, FeedbackClient, QueryClient, SourceClient,
+    decode_execute_sql_response, default_workspace,
 };
 use rmcp::{
     ErrorData, ServerHandler,
@@ -26,21 +21,19 @@ use rmcp::{
 };
 use serde::Serialize;
 use serde_json::{Map, Value};
-use std::sync::Arc;
 use tonic::Request;
 
 use crate::{
     McpOptions,
-    result_store::{ResultStore, ResultStoreError},
+    result_handles::{ResultHandles, query_result_status},
     surface::{
-        CatalogToolKind, RESULT_GET_MAX_LIMIT, ToolDescriptionContext, build_tool_result,
-        describe_table_arguments, describe_table_tool, describe_table_value, feedback_tool,
-        guide_resource, guide_resource_content, initial_instructions, list_catalog_arguments,
-        list_catalog_tool, list_catalog_value, list_columns_arguments, list_columns_tool,
-        list_columns_value, required_string_argument, result_get_arguments, result_get_tool,
-        search_catalog_arguments, search_catalog_tool, search_catalog_value, sql_tool,
-        status_to_error_data, tables_resource, tables_resource_content, tool_error_from_status,
-        tool_error_result,
+        CatalogToolKind, ToolDescriptionContext, build_tool_result, describe_table_arguments,
+        describe_table_tool, describe_table_value, feedback_tool, guide_resource,
+        guide_resource_content, initial_instructions, list_catalog_arguments, list_catalog_tool,
+        list_catalog_value, list_columns_arguments, list_columns_tool, list_columns_value,
+        required_string_argument, result_get_arguments, result_get_tool, search_catalog_arguments,
+        search_catalog_tool, search_catalog_value, sql_tool, status_to_error_data, tables_resource,
+        tables_resource_content, tool_error_from_status, tool_error_result,
     },
     telemetry,
 };
@@ -50,11 +43,6 @@ const LIST_CATALOG_COUNT_LIMIT: u32 = 1;
 const CATALOG_KIND_ALL: ProtoCatalogItemKind = ProtoCatalogItemKind::Unspecified;
 const CATALOG_KIND_TABLE: ProtoCatalogItemKind = ProtoCatalogItemKind::Table;
 const CATALOG_KIND_TABLE_FUNCTION: ProtoCatalogItemKind = ProtoCatalogItemKind::TableFunction;
-const SQL_PREVIEW_ROWS: usize = 20;
-const SQL_INLINE_BYTE_CHECK_MAX_ROWS: usize = 100;
-const SQL_INLINE_MAX_BYTES: usize = 8192;
-const LARGE_RESULT_GUIDANCE_MIN_ROWS: usize = 1_000;
-const LARGE_RESULT_GUIDANCE: &str = "Result is large; answer from row_count or rerun the SQL with filters or aggregates instead of paging every row. If raw rows are required, call result_get with limit 500 and a columns projection.";
 
 enum ToolCallOutcome {
     Success(Value),
@@ -62,63 +50,6 @@ enum ToolCallOutcome {
         operation: &'static str,
         status: tonic::Status,
     },
-}
-
-#[derive(Serialize)]
-struct SqlRowsValue {
-    rows: Vec<Value>,
-}
-
-#[derive(Serialize)]
-struct SqlHandledValue {
-    result_id: String,
-    row_count: usize,
-    column_count: usize,
-    columns: Vec<ColumnSummary>,
-    preview: ResultPreviewValue,
-    next_call: Option<NextCallValue>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    guidance: Option<&'static str>,
-}
-
-#[derive(Serialize)]
-struct SqlPreviewOnlyValue {
-    preview_only: bool,
-    row_count: usize,
-    column_count: usize,
-    columns: Vec<ColumnSummary>,
-    preview: ResultPreviewValue,
-    warning: &'static str,
-}
-
-#[derive(Serialize)]
-struct ResultPreviewValue {
-    offset: usize,
-    limit: usize,
-    has_more: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    next_offset: Option<usize>,
-    rows: Vec<Value>,
-}
-
-#[derive(Serialize)]
-struct NextCallValue {
-    tool: &'static str,
-    arguments: NextCallArgumentsValue,
-}
-
-#[derive(Serialize)]
-struct NextCallArgumentsValue {
-    result_id: String,
-    offset: usize,
-    limit: usize,
-}
-
-#[derive(Serialize)]
-struct ResultGetValue {
-    result_id: String,
-    #[serde(flatten)]
-    page: ResultPage,
 }
 
 #[derive(Serialize)]
@@ -147,7 +78,7 @@ pub(crate) struct CoralMcpServer {
     catalog: CatalogClient,
     query: QueryClient,
     feedback: FeedbackClient,
-    result_store: ResultStore,
+    result_handles: ResultHandles,
     options: McpOptions,
 }
 
@@ -158,7 +89,7 @@ impl CoralMcpServer {
             catalog: app.catalog_client(),
             query: app.query_client(),
             feedback: app.feedback_client(),
-            result_store: ResultStore::new(),
+            result_handles: ResultHandles::new(),
             options,
         }
     }
@@ -296,60 +227,7 @@ impl CoralMcpServer {
 
     async fn execute_sql_value(&self, sql: &str) -> Result<Value, tonic::Status> {
         let result = self.execute_sql_result(sql).await?;
-        if let Some(value) = inline_sql_value_if_small(&result)? {
-            return Ok(value);
-        }
-
-        // JSON object rows collapse duplicate column names and result_get
-        // projects columns by name, so duplicate-named results cannot be
-        // paged through a handle. Keep the legacy inline shape for them.
-        if has_duplicate_column_names(&result) {
-            let rows = batches_to_json_rows_json_safe_numbers(result.batches())
-                .map_err(|error| query_result_status(&error))?;
-            return serialize_tool_value(SqlRowsValue { rows });
-        }
-
-        let result = Arc::new(result);
-        let preview = preview_page(&result)?;
-        let columns = schema_summary(&result);
-        let estimated_bytes = result_estimated_bytes(&result);
-        match self
-            .result_store
-            .insert(Arc::clone(&result), estimated_bytes)
-        {
-            Ok(result_id) => {
-                // Advertise the maximum page size: agents tend to copy these
-                // arguments verbatim, and small pages multiply round trips.
-                let next_call = preview.next_offset.map(|offset| NextCallValue {
-                    tool: "result_get",
-                    arguments: NextCallArgumentsValue {
-                        result_id: result_id.clone(),
-                        offset,
-                        limit: RESULT_GET_MAX_LIMIT,
-                    },
-                });
-                let guidance = (result.row_count() >= LARGE_RESULT_GUIDANCE_MIN_ROWS)
-                    .then_some(LARGE_RESULT_GUIDANCE);
-                serialize_tool_value(SqlHandledValue {
-                    result_id,
-                    row_count: result.row_count(),
-                    column_count: columns.len(),
-                    columns,
-                    preview,
-                    next_call,
-                    guidance,
-                })
-            }
-            Err(ResultStoreError::TooLarge { .. }) => serialize_tool_value(SqlPreviewOnlyValue {
-                preview_only: true,
-                row_count: result.row_count(),
-                column_count: columns.len(),
-                columns,
-                preview,
-                warning: "Result exceeded the in-memory handle limit; rerun the SQL with LIMIT, filters, or a smaller column set.",
-            }),
-            Err(error) => Err(result_store_status(&error)),
-        }
+        self.result_handles.sql_value(result)
     }
 
     fn result_get_tool_result(
@@ -357,24 +235,12 @@ impl CoralMcpServer {
         request_arguments: Option<&Map<String, Value>>,
     ) -> Result<ToolCallOutcome, ErrorData> {
         let arguments = result_get_arguments(request_arguments)?;
-        let result = self
-            .result_store
-            .get(&arguments.result_id)
-            .map_err(|error| result_store_error_data(&error))?;
-        let page = slice_result(
-            &result,
-            ResultSliceRequest {
-                offset: arguments.offset,
-                limit: arguments.limit,
-                columns: arguments.columns.as_deref(),
-            },
-        )
-        .map_err(result_page_error_data)?;
-        let value = serde_json::to_value(ResultGetValue {
-            result_id: arguments.result_id,
-            page,
-        })
-        .map_err(|error| ErrorData::internal_error(error.to_string(), None))?;
+        let value = self.result_handles.page_value(
+            arguments.result_id,
+            arguments.offset,
+            arguments.limit,
+            arguments.columns.as_deref(),
+        )?;
         Ok(ToolCallOutcome::Success(value))
     }
 
@@ -734,71 +600,6 @@ fn finish_tool_call(
             telemetry::record_protocol_error(span, &error);
             Err(error)
         }
-    }
-}
-
-fn inline_sql_value_if_small(
-    result: &CollectedQueryResult,
-) -> Result<Option<Value>, tonic::Status> {
-    if result.row_count() > SQL_INLINE_BYTE_CHECK_MAX_ROWS {
-        return Ok(None);
-    }
-    let rows = batches_to_json_rows_json_safe_numbers(result.batches())
-        .map_err(|error| query_result_status(&error))?;
-    let value = serialize_tool_value(SqlRowsValue { rows })?;
-    let compact_len = serde_json::to_string(&value)
-        .map_err(|error| tonic::Status::internal(error.to_string()))?
-        .len();
-    let preview_would_include_full_result = result.row_count() <= SQL_PREVIEW_ROWS;
-    if preview_would_include_full_result || compact_len <= SQL_INLINE_MAX_BYTES {
-        Ok(Some(value))
-    } else {
-        Ok(None)
-    }
-}
-
-fn preview_page(result: &CollectedQueryResult) -> Result<ResultPreviewValue, tonic::Status> {
-    let page = slice_result(
-        result,
-        ResultSliceRequest {
-            offset: 0,
-            limit: SQL_PREVIEW_ROWS,
-            columns: None,
-        },
-    )
-    .map_err(|error| query_result_status(&error))?;
-    Ok(ResultPreviewValue {
-        offset: page.offset,
-        limit: page.limit,
-        has_more: page.has_more,
-        next_offset: page.next_offset,
-        rows: page.rows,
-    })
-}
-
-fn query_result_status(error: &QueryResultError) -> tonic::Status {
-    tonic::Status::internal(error.to_string())
-}
-
-fn result_store_error_data(error: &ResultStoreError) -> ErrorData {
-    match error {
-        ResultStoreError::NotFound(_) | ResultStoreError::Expired(_) => {
-            ErrorData::invalid_params(error.to_string(), None)
-        }
-        ResultStoreError::TooLarge { .. } | ResultStoreError::Unavailable => {
-            ErrorData::internal_error(error.to_string(), None)
-        }
-    }
-}
-
-fn result_store_status(error: &ResultStoreError) -> tonic::Status {
-    tonic::Status::internal(error.to_string())
-}
-
-fn result_page_error_data(error: QueryResultError) -> ErrorData {
-    match error {
-        QueryResultError::InvalidSliceRequest(message) => ErrorData::invalid_params(message, None),
-        other => ErrorData::internal_error(other.to_string(), None),
     }
 }
